feat: add development environment support with dynamic logo and version display
This commit introduces a new hook, useEnvironment, to determine if the application is running in a development environment. The AppSidebar and DashboardLayout components are updated to conditionally display a development logo and append "DEV" to the application name and version when in development mode. Additionally, a new logodev.png file is added for use in development.